Fixed error with packing lat,, lon. 32 bit int now used to pack
I tested vigorously to ensure packing and unpacking were working

+++ src/starvationevasion/sim/Model.java
@@ -334,7 +334,7 @@ public class Model
     return LandTile.PACKED_CROP_RATINGS;
   }
 
-  public short[] getPackedTileCoordinates()
+  public int[] getPackedTileCoordinates()
   {
     return LandTile.PACKED_TILE_COORDINATES;
   }
